@@ -10,13 +10,19 @@ description: A simple mobile app for tracking activities.
 version: 1.2.1+20230814
 
 environment:
-  sdk: '>=3.2.3 <4.0.0'
+  # TODO: Update to 3.7.0 when https://github.com/dart-lang/dart_style/commit/71421b328391a6ed9c1a9e7229c394d9ed613d77 is released.
+  sdk: ^3.6.0
+  flutter: ">=1.17.0"
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
+
+  adair_flutter_lib:
+    # ignore: invalid_dependency
+    path: ../../adair-flutter-lib
 
   community_charts_flutter: ^1.0.4  # https://pub.dev/packages/community_charts_flutter
   collection: ^1.19.1               # https://pub.dev/packages/collection (Dart)
@@ -40,6 +46,7 @@ dependencies:
   firebase_crashlytics: ^4.3.3      # https://pub.dev/packages/firebase_crashlytics
   firebase_core: ^3.12.0            # https://pub.dev/packages/firebase_core
   firebase_analytics: ^11.4.3       # https://pub.dev/packages/firebase_analytics
+  timezone: ^0.10.0                 # https://pub.dev/packages/timezone
 
 dev_dependencies:
   flutter_test:
