@@ -1,17 +1,13 @@
+import 'package:adair_flutter_lib/wrappers/io_wrapper.dart';
 import 'package:mobile/http_wrapper.dart';
-import 'package:mobile/io_wrapper.dart';
 import 'package:mobile/package_info_wrapper.dart';
 import 'package:mobile/preferences_manager.dart';
 import 'package:mobile/database/data_manager.dart';
-import 'package:mobile/properties_manager.dart';
 
 import 'device_info_wrapper.dart';
 
 class AppManager {
-  PropertiesManager? _propertiesManager;
-
   DeviceInfoWrapper? _deviceInfoWrapper;
-  IoWrapper? _ioWrapper;
   PackageInfoWrapper? _packageInfoWrapper;
   HttpWrapper? _httpWrapper;
 
@@ -19,19 +15,9 @@ class AppManager {
 
   PreferencesManager get preferencesManager => PreferencesManager.get;
 
-  PropertiesManager get propertiesManager {
-    _propertiesManager ??= PropertiesManager();
-    return _propertiesManager!;
-  }
-
   DeviceInfoWrapper get deviceInfoWrapper {
     _deviceInfoWrapper ??= DeviceInfoWrapper();
     return _deviceInfoWrapper!;
-  }
-
-  IoWrapper get ioWrapper {
-    _ioWrapper ??= IoWrapper();
-    return _ioWrapper!;
   }
 
   PackageInfoWrapper get packageInfoWrapper {
