@@ -1,6 +1,6 @@
+import 'package:adair_flutter_lib/res/theme.dart';
 import 'package:community_charts_flutter/community_charts_flutter.dart';
 import 'package:flutter/material.dart';
-import 'package:mobile/res/theme.dart';
 
 RenderSpec<num>? defaultChartRenderSpec(BuildContext context) {
   return GridlineRendererSpec(
