@@ -32,43 +32,41 @@ class ActivityPickerState extends State<ActivityPicker> {
   @override
   Widget build(BuildContext context) {
     return ActivitiesBuilder(
-        app: widget.app,
-        builder: (BuildContext context, List<Activity> activities) {
-          _allActivitiesActivity = ActivityBuilder(
-            Strings.of(context).activityDropdownAllActivities,
-          ).build;
+      app: widget.app,
+      builder: (BuildContext context, List<Activity> activities) {
+        _allActivitiesActivity = ActivityBuilder(
+          Strings.of(context).activityDropdownAllActivities,
+        ).build;
 
-          return ListPicker<Activity>(
-            allowsMultiSelect: true,
-            initialValues: widget.initialActivities.isEmpty
-                ? {_allActivitiesActivity}
-                : widget.initialActivities,
-            onChanged: (Set<Activity> newActivities) {
-              if (newActivities.first == _allActivitiesActivity) {
-                // Invoke the callback with null if "All activities" was picked.
-                widget.onPickedActivitiesChanged({});
-              } else {
-                widget.onPickedActivitiesChanged(newActivities);
-              }
-            },
-            allItem: _buildItem(_allActivitiesActivity),
-            items: activities.map((activity) => _buildItem(activity)).toList(),
-            titleBuilder: (Set<Activity> selectedActivities) {
-              return Text(
-                selectedActivities
-                    .map((activity) => activity.name)
-                    .toList()
-                    .join(", "),
-              );
-            },
-          );
-        });
+        return ListPicker<Activity>(
+          allowsMultiSelect: true,
+          initialValues: widget.initialActivities.isEmpty
+              ? {_allActivitiesActivity}
+              : widget.initialActivities,
+          onChanged: (Set<Activity> newActivities) {
+            if (newActivities.first == _allActivitiesActivity) {
+              // Invoke the callback with null if "All activities" was picked.
+              widget.onPickedActivitiesChanged({});
+            } else {
+              widget.onPickedActivitiesChanged(newActivities);
+            }
+          },
+          allItem: _buildItem(_allActivitiesActivity),
+          items: activities.map((activity) => _buildItem(activity)).toList(),
+          titleBuilder: (Set<Activity> selectedActivities) {
+            return Text(
+              selectedActivities
+                  .map((activity) => activity.name)
+                  .toList()
+                  .join(", "),
+            );
+          },
+        );
+      },
+    );
   }
 
   ListPickerItem<Activity> _buildItem(Activity activity) {
-    return ListPickerItem<Activity>(
-      value: activity,
-      title: activity.name,
-    );
+    return ListPickerItem<Activity>(value: activity, title: activity.name);
   }
 }
