@@ -1,7 +1,7 @@
 import 'dart:async';
 
+import 'package:adair_flutter_lib/widgets/empty.dart';
 import 'package:flutter/material.dart';
-import 'package:mobile/widgets/widget.dart';
 
 /// A utility [Widget] capable of listening to multiple [Stream]s.
 ///
@@ -68,8 +68,9 @@ class FutureListenerState extends State<FutureListener> {
     super.initState();
 
     for (var stream in widget.streams) {
-      _onUpdateEvents
-          .add(stream.listen((newValue) => setState(() => _updateFutures())));
+      _onUpdateEvents.add(
+        stream.listen((newValue) => setState(() => _updateFutures())),
+      );
     }
 
     _updateFutures();
@@ -90,7 +91,7 @@ class FutureListenerState extends State<FutureListener> {
       builder: (BuildContext context, AsyncSnapshot snapshot) {
         if (!snapshot.hasData) {
           if (widget.initialValues.isEmpty) {
-            return Empty();
+            return const Empty();
           }
 
           return _build(
