@@ -1,12 +1,12 @@
+import 'package:adair_flutter_lib/res/dimen.dart';
+import 'package:adair_flutter_lib/utils/dialog.dart';
+import 'package:adair_flutter_lib/widgets/empty.dart';
 import 'package:flutter/material.dart';
 import 'package:mobile/app_manager.dart';
 import 'package:mobile/i18n/strings.dart';
 import 'package:mobile/model/session.dart';
-import 'package:mobile/res/dimen.dart';
-import 'package:mobile/utils/dialog_utils.dart';
 import 'package:mobile/widgets/list_item.dart';
 import 'package:mobile/widgets/text.dart';
-import 'package:mobile/widgets/widget.dart';
 
 typedef OnTapSessionListTile = Function(Session);
 
@@ -57,7 +57,7 @@ class SessionListTile extends StatelessWidget {
             },
           ),
         ),
-        _hasDivider ? const Divider(height: 1) : Empty(),
+        _hasDivider ? const Divider(height: 1) : const Empty(),
       ],
     );
   }
