@@ -1,12 +1,14 @@
+import 'package:adair_flutter_lib/managers/time_manager.dart';
+import 'package:adair_flutter_lib/utils/date_range.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mobile/database/data_manager.dart';
 import 'package:mobile/model/activity.dart';
 import 'package:mobile/model/session.dart';
-import 'package:mobile/utils/date_time_utils.dart';
+import 'package:mobile/utils/duration.dart';
 import 'package:mockito/mockito.dart';
 
 import '../mocks/mocks.mocks.dart';
-import '../test_utils.dart';
+import '../stubbed_managers.dart';
 
 void main() {
   late MockAppManager appManager;
@@ -14,6 +16,7 @@ void main() {
   late MockPreferencesManager preferencesManager;
 
   setUp(() async {
+    await StubbedManagers.create(); // Initializes TimeManager.
     appManager = MockAppManager();
 
     var batch = MockBatch();
@@ -24,35 +27,49 @@ void main() {
 
     preferencesManager = MockPreferencesManager();
     when(appManager.preferencesManager).thenReturn(preferencesManager);
-    when(preferencesManager.largestDurationUnit).thenReturn(DurationUnit.days);
-    when(preferencesManager.homeDateRange)
-        .thenReturn(DisplayDateRange.allDates);
+    when(
+      preferencesManager.largestDurationUnit,
+    ).thenReturn(AppDurationUnit.days);
+    when(
+      preferencesManager.homeDateRange,
+    ).thenReturn(DisplayDateRange.allDates);
 
     DataManager.suicide();
     await DataManager.get.init(appManager, database);
   });
 
   stubActivities(List<Map<String, dynamic>> result) {
-    when(database.rawQuery("SELECT * FROM activity"))
-        .thenAnswer((_) async => result);
+    when(
+      database.rawQuery("SELECT * FROM activity"),
+    ).thenAnswer((_) async => result);
   }
 
   group("getSummarizedActivities", () {
-    stubOverlappingSessions(String activityId, DateRange dateRange,
-        List<Map<String, dynamic>> result) {
-      when(database.rawQuery("""
+    stubOverlappingSessions(
+      String activityId,
+      DateRange dateRange,
+      List<Map<String, dynamic>> result,
+    ) {
+      when(
+        database.rawQuery(
+          """
           SELECT * FROM session
             WHERE activity_id = ?
             AND start_timestamp < ?
             AND (end_timestamp IS NULL OR end_timestamp > ?)
             AND is_banked = 0
             ORDER BY start_timestamp
-          """, [activityId, dateRange.endMs, dateRange.startMs]))
-          .thenAnswer((_) async => result);
+          """,
+          [activityId, dateRange.endMs, dateRange.startMs],
+        ),
+      ).thenAnswer((_) async => result);
     }
 
     Map<String, dynamic> buildSession(
-        String activityId, DateTime start, DateTime end) {
+      String activityId,
+      DateTime start,
+      DateTime end,
+    ) {
       return (SessionBuilder(activityId)
             ..startTimestamp = start.millisecondsSinceEpoch
             ..endTimestamp = end.millisecondsSinceEpoch)
@@ -67,24 +84,27 @@ void main() {
       required int expectedLength,
       required Duration expectedDuration,
     }) async {
-      DisplayDateRange dateRange = stubDateRange(DateRange(
-        startDate: startDate,
-        endDate: endDate,
-      ));
+      var dateRange = DisplayDateRange.dateRange(
+        DateRange(
+          startDate: TimeManager.get.dateTimeToTz(startDate),
+          endDate: TimeManager.get.dateTimeToTz(endDate),
+        ),
+      );
 
       Activity activity = ActivityBuilder("").build;
 
       stubActivities([activity.toMap()]);
       stubOverlappingSessions(
-          activity.id,
-          dateRange.value,
-          sessionRangeList.map((DateRange dateRange) {
-            return buildSession(
-              activity.id,
-              dateRange.startDate,
-              dateRange.endDate,
-            );
-          }).toList());
+        activity.id,
+        dateRange.value,
+        sessionRangeList.map((DateRange dateRange) {
+          return buildSession(
+            activity.id,
+            dateRange.startDate,
+            dateRange.endDate,
+          );
+        }).toList(),
+      );
 
       var result = await DataManager.get.getSummarizedActivities(dateRange);
 
@@ -96,10 +116,12 @@ void main() {
 
     test("No activities", () async {
       stubActivities([]);
-      DisplayDateRange dateRange = stubDateRange(DateRange(
-        startDate: DateTime.now(),
-        endDate: DateTime.now(),
-      ));
+      var dateRange = DisplayDateRange.dateRange(
+        DateRange(
+          startDate: TimeManager.get.now(),
+          endDate: TimeManager.get.now(),
+        ),
+      );
       var result = await DataManager.get.getSummarizedActivities(dateRange);
       expect(result.activities, isEmpty);
       expect(result.longestSession, isNull);
@@ -107,10 +129,12 @@ void main() {
     });
 
     test("Activities provided as parameter", () async {
-      DisplayDateRange dateRange = stubDateRange(DateRange(
-        startDate: DateTime(2018, 1, 1),
-        endDate: DateTime(2018, 2, 1),
-      ));
+      var dateRange = DisplayDateRange.dateRange(
+        DateRange(
+          startDate: TimeManager.get.dateTimeFromValues(2018, 1, 1),
+          endDate: TimeManager.get.dateTimeFromValues(2018, 2, 1),
+        ),
+      );
 
       Activity activity = ActivityBuilder("").build;
 
@@ -118,8 +142,8 @@ void main() {
       stubOverlappingSessions(activity.id, dateRange.value, [
         buildSession(
           activity.id,
-          DateTime(2018, 1, 15, 5),
-          DateTime(2018, 1, 15, 7),
+          TimeManager.get.dateTimeFromValues(2018, 1, 15, 5),
+          TimeManager.get.dateTimeFromValues(2018, 1, 15, 7),
         ),
       ]);
 
@@ -148,12 +172,12 @@ void main() {
 
     test("Session start outside range, session end inside range", () async {
       await assertSummarizedActivities(
-        startDate: DateTime(2018, 1, 15, 4, 30),
-        endDate: DateTime(2018, 1, 16, 4, 30),
+        startDate: TimeManager.get.dateTimeFromValues(2018, 1, 15, 4, 30),
+        endDate: TimeManager.get.dateTimeFromValues(2018, 1, 16, 4, 30),
         sessionRangeList: [
           DateRange(
-            startDate: DateTime(2018, 1, 15, 2, 30),
-            endDate: DateTime(2018, 1, 15, 7, 30),
+            startDate: TimeManager.get.dateTimeFromValues(2018, 1, 15, 2, 30),
+            endDate: TimeManager.get.dateTimeFromValues(2018, 1, 15, 7, 30),
           ),
         ],
         expectedLength: 1,
@@ -163,12 +187,12 @@ void main() {
 
     test("Session start inside range, session end outside range", () async {
       await assertSummarizedActivities(
-        startDate: DateTime(2018, 1, 15, 4, 30),
-        endDate: DateTime(2018, 1, 16, 4, 30),
+        startDate: TimeManager.get.dateTimeFromValues(2018, 1, 15, 4, 30),
+        endDate: TimeManager.get.dateTimeFromValues(2018, 1, 16, 4, 30),
         sessionRangeList: [
           DateRange(
-            startDate: DateTime(2018, 1, 16, 2, 30),
-            endDate: DateTime(2018, 1, 16, 7, 30),
+            startDate: TimeManager.get.dateTimeFromValues(2018, 1, 16, 2, 30),
+            endDate: TimeManager.get.dateTimeFromValues(2018, 1, 16, 7, 30),
           ),
         ],
         expectedLength: 1,
@@ -178,12 +202,12 @@ void main() {
 
     test("Session start outside range, session end outside range", () async {
       await assertSummarizedActivities(
-        startDate: DateTime(2018, 1, 15, 4, 30),
-        endDate: DateTime(2018, 1, 16, 4, 30),
+        startDate: TimeManager.get.dateTimeFromValues(2018, 1, 15, 4, 30),
+        endDate: TimeManager.get.dateTimeFromValues(2018, 1, 16, 4, 30),
         sessionRangeList: [
           DateRange(
-            startDate: DateTime(2018, 1, 15, 2, 30),
-            endDate: DateTime(2018, 1, 16, 7, 30),
+            startDate: TimeManager.get.dateTimeFromValues(2018, 1, 15, 2, 30),
+            endDate: TimeManager.get.dateTimeFromValues(2018, 1, 16, 7, 30),
           ),
         ],
         expectedLength: 1,
@@ -193,12 +217,12 @@ void main() {
 
     test("Session start inside range, session end inside range", () async {
       await assertSummarizedActivities(
-        startDate: DateTime(2018, 1, 15, 4, 30),
-        endDate: DateTime(2018, 1, 16, 4, 30),
+        startDate: TimeManager.get.dateTimeFromValues(2018, 1, 15, 4, 30),
+        endDate: TimeManager.get.dateTimeFromValues(2018, 1, 16, 4, 30),
         sessionRangeList: [
           DateRange(
-            startDate: DateTime(2018, 1, 15, 8, 30),
-            endDate: DateTime(2018, 1, 15, 20, 30),
+            startDate: TimeManager.get.dateTimeFromValues(2018, 1, 15, 8, 30),
+            endDate: TimeManager.get.dateTimeFromValues(2018, 1, 15, 20, 30),
           ),
         ],
         expectedLength: 1,
@@ -208,12 +232,12 @@ void main() {
 
     test("Session start outside range, session end == range start", () async {
       await assertSummarizedActivities(
-        startDate: DateTime(2018, 1, 15, 4, 30),
-        endDate: DateTime(2018, 1, 16, 4, 30),
+        startDate: TimeManager.get.dateTimeFromValues(2018, 1, 15, 4, 30),
+        endDate: TimeManager.get.dateTimeFromValues(2018, 1, 16, 4, 30),
         sessionRangeList: [
           DateRange(
-            startDate: DateTime(2018, 1, 14, 8, 30),
-            endDate: DateTime(2018, 1, 15, 4, 30),
+            startDate: TimeManager.get.dateTimeFromValues(2018, 1, 14, 8, 30),
+            endDate: TimeManager.get.dateTimeFromValues(2018, 1, 15, 4, 30),
           ),
         ],
         expectedLength: 1,
@@ -223,12 +247,12 @@ void main() {
 
     test("Session start == range end, session end outside range", () async {
       await assertSummarizedActivities(
-        startDate: DateTime(2018, 1, 15, 4, 30),
-        endDate: DateTime(2018, 1, 16, 4, 30),
+        startDate: TimeManager.get.dateTimeFromValues(2018, 1, 15, 4, 30),
+        endDate: TimeManager.get.dateTimeFromValues(2018, 1, 16, 4, 30),
         sessionRangeList: [
           DateRange(
-            startDate: DateTime(2018, 1, 16, 4, 30),
-            endDate: DateTime(2018, 1, 18, 4, 30),
+            startDate: TimeManager.get.dateTimeFromValues(2018, 1, 16, 4, 30),
+            endDate: TimeManager.get.dateTimeFromValues(2018, 1, 18, 4, 30),
           ),
         ],
         expectedLength: 1,
@@ -237,10 +261,12 @@ void main() {
     });
 
     test("Combination of all with multiple activities", () async {
-      DisplayDateRange dateRange = stubDateRange(DateRange(
-        startDate: DateTime(2018, 1, 1),
-        endDate: DateTime(2018, 2, 1),
-      ));
+      var dateRange = DisplayDateRange.dateRange(
+        DateRange(
+          startDate: TimeManager.get.dateTimeFromValues(2018, 1, 1),
+          endDate: TimeManager.get.dateTimeFromValues(2018, 2, 1),
+        ),
+      );
 
       List<Activity> activities = [
         ActivityBuilder("Activity 1").build,
@@ -251,57 +277,58 @@ void main() {
       ];
 
       stubActivities(
-          activities.map((Activity activity) => activity.toMap()).toList());
+        activities.map((Activity activity) => activity.toMap()).toList(),
+      );
 
       stubOverlappingSessions(activities[0].id, dateRange.value, [
         buildSession(
           activities[0].id,
-          DateTime(2017, 12, 31, 22),
-          DateTime(2018, 1, 1, 4),
+          TimeManager.get.dateTimeFromValues(2017, 12, 31, 22),
+          TimeManager.get.dateTimeFromValues(2018, 1, 1, 4),
         ), // Expected 4 hours
         buildSession(
           activities[0].id,
-          DateTime(2018, 1, 14, 12),
-          DateTime(2018, 1, 14, 15),
+          TimeManager.get.dateTimeFromValues(2018, 1, 14, 12),
+          TimeManager.get.dateTimeFromValues(2018, 1, 14, 15),
         ), // Expected 3 hours
         buildSession(
           activities[0].id,
-          DateTime(2018, 1, 31, 15),
-          DateTime(2018, 2, 1, 15),
+          TimeManager.get.dateTimeFromValues(2018, 1, 31, 15),
+          TimeManager.get.dateTimeFromValues(2018, 2, 1, 15),
         ), // Expected 9 hours
       ]);
 
       stubOverlappingSessions(activities[1].id, dateRange.value, [
         buildSession(
           activities[1].id,
-          DateTime(2018, 1, 9, 9),
-          DateTime(2018, 1, 9, 17),
+          TimeManager.get.dateTimeFromValues(2018, 1, 9, 9),
+          TimeManager.get.dateTimeFromValues(2018, 1, 9, 17),
         ), // Expected 8 hours
       ]);
 
       stubOverlappingSessions(activities[2].id, dateRange.value, [
         buildSession(
           activities[2].id,
-          DateTime(2018, 1, 1),
-          DateTime(2018, 1, 1, 15),
+          TimeManager.get.dateTimeFromValues(2018, 1, 1),
+          TimeManager.get.dateTimeFromValues(2018, 1, 1, 15),
         ), // Expected 15 hours
         buildSession(
           activities[2].id,
-          DateTime(2018, 1, 20, 12),
-          DateTime(2018, 1, 20, 24),
+          TimeManager.get.dateTimeFromValues(2018, 1, 20, 12),
+          TimeManager.get.dateTimeFromValues(2018, 1, 20, 24),
         ), // Expected 12 hours
         buildSession(
           activities[2].id,
-          DateTime(2018, 1, 31, 15),
-          DateTime(2018, 2, 1),
+          TimeManager.get.dateTimeFromValues(2018, 1, 31, 15),
+          TimeManager.get.dateTimeFromValues(2018, 2, 1),
         ), // Expected 9 hours
       ]);
 
       stubOverlappingSessions(activities[3].id, dateRange.value, [
         buildSession(
           activities[3].id,
-          DateTime(2018, 1, 1),
-          DateTime(2018, 2, 1),
+          TimeManager.get.dateTimeFromValues(2018, 1, 1),
+          TimeManager.get.dateTimeFromValues(2018, 2, 1),
         ), // Expected 31 days
       ]);
 
@@ -346,8 +373,9 @@ void main() {
   });
 
   test("getInProgressSession returns null", () async {
-    when(database.rawQuery(any, any))
-        .thenAnswer((_) => Future.value(List.empty()));
+    when(
+      database.rawQuery(any, any),
+    ).thenAnswer((_) => Future.value(List.empty()));
     expect(await DataManager.get.inProgressSession("id"), isNull);
   });
 
